requirements.txt
backoff==2.1.2
click==8.1.3
Flask==2.2.2
flask-paginate==2022.1.8
gql==3.4.0
graphql-core==3.2.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
python-dotenv==0.20.0
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.8.1